+++ src/main/java/africa/semicolon/playlist/exceptions/PlaylistNotFoundException.java
@@ -1,0 +1,4 @@
+package africa.semicolon.playlist.exceptions;
+
+public class PlaylistNotFoundException {
+}
