+++ src/main/java/africa/semicolon/playlist/playlist/dto/CreatePlaylistReq.java
@@ -1,0 +1,2 @@
+package africa.semicolon.playlist.playlist.dto;public class CreatePlaylistReq {
+}
