+++ src/main/java/africa/semicolon/playlist/playlist/dto/CreatePlaylistResponse.java
@@ -1,0 +1,2 @@
+package africa.semicolon.playlist.playlist.dto;public class CreatePlaylistResponse {
+}
