+++ src/main/java/africa/semicolon/playlist/playlist/dto/UploadPlaylistImageResponse.java
@@ -1,0 +1,6 @@
+package africa.semicolon.playlist.playlist.dto;
+
+public class UploadPlaylistResponse {
+
+    private
+}
