+++ src/main/java/africa/semicolon/playlist/playlist/service/PlaylistServiceImpl.java
@@ -1,0 +1,2 @@
+package africa.semicolon.playlist.playlist.service;public class PlaylistServiceImpl {
+}
