+++ src/main/java/africa/semicolon/playlist/playlistSong/service/PlaylistUserServiceImpl.java
@@ -1,0 +1,2 @@
+package africa.semicolon.playlist.playlistSong.service;public class PlaylistSongServiceImpl {
+}
