+++ src/main/java/africa/semicolon/playlist/song/dto/CreateSongResponse.java
@@ -1,0 +1,2 @@
+package africa.semicolon.playlist.song.dto;public class CreateSongResponse {
+}
