--- src/main/java/africa/semicolon/playlist/user/UserEntity.java
@@ -1,27 +1,0 @@
-package africa.semicolon.playlist.user;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.*;
-
-@Entity
-@Setter
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-
-public class UserEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long songId;
-    private String firstName;
-    private String lastName;
-    private String emailAddress;
-    private String password;
-    @JsonFormat(pattern = "yyyy-MM-dd")
-    private String dateJoined;
-}
